requirements.txt
app-store-scraper==0.3.5
asgiref==3.4.1
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
fastapi==0.68.1
google-play-scraper==1.0.1
h11==0.12.0
idna==2.10
pydantic==1.8.2
PyYAML==5.4.1
requests==2.23.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.25.11
uvicorn==0.15.0
